Moved window() from local/util/window.py to modelgen/window.py. Also removed ModelGenWindow named tuples from the pipeline.
